feat: Add axios npm dependency for HTTP requests
This commit adds the axios npm dependency to the project, allowing for easier and more efficient HTTP requests. Axios is a popular library for making HTTP requests in JavaScript, and it will be used in the code to fetch data from the API endpoint. This addition will enhance the functionality of the application by enabling seamless communication with the server.